Select u-boot based on boot/loader/uEnv.txt
This change is backward compatible with the previous
ostree releases where it used to check if /boot/uEnv.txt
is a symlink to detect if it should initialize u-boot
handler code.

Many embedded device vendors expect to find uEnv.txt in
the top level directory, it is often hardcoded when building
u-boot and is difficult to change later on. With this patch
adding ostree support on such boards becomes simpler.

Now to add ostree support on such boards you simply add a custom
logic in uEnv.txt that loads ostree env from /loader/uEnv.txt

This patch mitigates the behaviour described in:

https://bugzilla.gnome.org/show_bug.cgi?id=755787