admin: Add a `dup` command
I'm trying to improve the developer experience on OSTree-managed
systems, and I had an epiphany the other day - there's no reason we
have to be absolutely against mutating the current rootfs live.  The
key should be making it easy to rollback/reset to a known good state.

This new `dup` command isn't a full solution by any means, but it's
a step towards a workflow that would look like this:

```
ostree admin dup          # clone the current root
ostree admin dev-unlock   # remove the ro bind mount, and add rofiles-fuse
ostree admin dev-reset    # not sure of the name...but basically resync
                          # files in the current root from a checkout
```