Allow updating files on the boot partition
Until now OSTree copied only the vmlinuz and initramfs
binaries to the boot/ directory (which in some setups
might be on a separate partition). This patch adds
support for copying other files from the deployment's
/boot directory to the real /boot.

How this works:

Ignore subdirectories, only files in root of the boot
directory are copied. There is overhead of copying files
to boot/, therefore the amount of files in boot/ should
be kept to the minimum and subdirectories shouldn't
really be necessary.

Files in the boot/ directory are updated only with major
releases, when kernel/initramfs bootcsum changes. Files
that require frequent updates should not be stored here.